Add cpu functionality when user rolls 1
Adds a line in roll action to check when you roll a 1, if the current
player is the user, then run the cpu function. Also adds the hold
argument for the hold button hold functions to align with the holdAction
function change to include the hold string when differentiating between
hold and switch.